lab6/src/api/controller/middleware/tracing.py
from opentelemetry import trace
from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
from starlette.requests import Request
from starlette.responses import Response

tracer = trace.get_tracer(__name__)


class TracingMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint) -> Response:
        with tracer.start_as_current_span(request.url.path):
            return await call_next(request)
