lab6/pyproject.toml
[tool.poetry]
authors = ["Ivan Krasnopolsky <i1.krasnopolskiy@gmail.com>"]
description = ""
name = "lab5"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
aio-pika = "^9.4.1"
alembic = "^1.13.1"
fastapi = "^0.111.0"
loguru = "^0.7.2"
mimesis = "^16.0.0"
mysqlclient = "^2.2.4"
opentelemetry-api = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-aio-pika = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-sdk = "^1.24.0"
pika = "^1.3.2"
pika-stubs = "^0.1.3"
prometheus-client = "^0.20.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
python = "^3.11"
python-logging-loki = "^0.3.1"
redis = "^5.0.4"
sqlalchemy = "^2.0.30"
structlog = "^24.1.0"
uvicorn = "^0.29.0"

[tool.poetry.group.dev.dependencies]
aiohttp = "^3.9.3"
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.1.1"
toml-sort = "^0.23.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
exclude = '(\.git|\.cache|\.venv|\.idea|__pycache__)'
line-length = 120

[tool.isort]
line_length = 120
profile = "black"
skip = ["**/__pycache__/**", ".cache", ".git", ".idea", ".venv"]
